Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1044931 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/order/src/org/ofbiz/order/shoppingcart/shipping/ShippingEstimateWrapper.java
@@ -116,7 +116,7 @@ public class ShippingEstimateWrapper {
     }
 
     public BigDecimal getShippingEstimate(GenericValue storeCarrierShipMethod) {
-        return (BigDecimal) shippingEstimates.get(storeCarrierShipMethod);
+        return shippingEstimates.get(storeCarrierShipMethod);
     }
 
 }
