+++ framework/base/src/org/ofbiz/base/util/TimeDuration.java
@@ -307,10 +307,10 @@ public class TimeDuration implements Serializable, Comparable<TimeDuration> {
         }
         long units = duration / 0x757B12C00L;
         int years = (int) units;
-        duration -= 0x757B12C00L * (long) years;
+        duration -= 0x757B12C00L * years;
         units = duration / 0x9CA41900L;
         int months = (int) units;
-        duration -= 0x9CA41900L * (long) months;
+        duration -= 0x9CA41900L * months;
         units = duration / 86400000;
         int days = (int) units;
         duration -= 86400000 * (long) days;
@@ -373,8 +373,8 @@ public class TimeDuration implements Serializable, Comparable<TimeDuration> {
      */
     public static long toLong(TimeDuration duration) {
         return
-        (0x757B12C00L * (long) duration.years) +
-        (0x9CA41900L * (long) duration.months) +
+        (0x757B12C00L * duration.years) +
+        (0x9CA41900L * duration.months) +
         (86400000 * (long) duration.days) +
         (3600000 * (long) duration.hours) +
         (60000 * (long) duration.minutes) +
