+++ framework/common/src/org/ofbiz/common/Captcha.java
@@ -107,7 +107,7 @@ public class Captcha {
                 // we can rotate it independently
                 int charWidth = fontMetrics.charWidth(characterToShow.charAt(i));
                 int charDim = Math.max(maxAdvance, fontHeight);
-                int halfCharDim = (int) (charDim / 2);
+                int halfCharDim = (charDim / 2);
 
                 BufferedImage charImage =
                         new BufferedImage(charDim, charDim, BufferedImage.TYPE_INT_ARGB);
@@ -121,10 +121,10 @@ public class Captcha {
 
                 int charX = (int) (0.5 * charDim - 0.5 * charWidth);
                 charGraphics.drawString("" + characterToShow.charAt(i), charX,
-                        (int) ((charDim - fontMetrics.getAscent()) / 2 + fontMetrics.getAscent()));
+                        ((charDim - fontMetrics.getAscent()) / 2 + fontMetrics.getAscent()));
 
                 float x = horizMargin + spacePerChar * (i) - charDim / 2.0f;
-                int y = (int) ((height - charDim) / 2);
+                int y = ((height - charDim) / 2);
 
                 g.drawImage(charImage, (int) x, y, charDim, charDim, null, null);
