+++ framework/service/src/org/ofbiz/service/semaphore/ServiceSemaphore.java
@@ -100,7 +100,7 @@ public class ServiceSemaphore {
                 }
             }
             if (timedOut) {
-                double waitTimeSec = ((double) (System.currentTimeMillis() - lockTime.getTime()) / 1000.0);
+                double waitTimeSec = ((System.currentTimeMillis() - lockTime.getTime()) / 1000.0);
                 String errMsg = "Service [" + model.name + "] with wait semaphore exceeded wait timeout, waited [" + waitTimeSec + "], wait started at " + lockTime;
                 Debug.logWarning(errMsg, module);
                 throw new SemaphoreWaitException(errMsg);
